# BackEnd/tests/_regression_snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RepsActivityCategoryCreate": {
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "RepsActivityCategoryCreate",
        "type": "object"
      },
      "RepsActivityCategoryRes": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_default": {
            "default": false,
            "title": "Is Default",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sort_order": {
            "default": 0,
            "title": "Sort Order",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "RepsActivityCategoryRes",
        "type": "object"
      },
      "RepsEntriesEnvelope": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/RepsEntryRow"
            },
            "title": "Entries",
            "type": "array"
          },
          "stats": {
            "$ref": "#/components/schemas/RepsStats"
          },
          "user": {
            "enum": [
              "Aviv2026",
              "Yarden2026"
            ],
            "title": "User",
            "type": "string"
          }
        },
        "required": [
          "user",
          "entries",
          "stats"
        ],
        "title": "RepsEntriesEnvelope",
        "type": "object"
      },
...
      "RepsLogCreate": {
        "description": "Payload accepted by `POST /reps/log`.\n\nThe backend overrides `created_at` server-side to a contemporaneous\nISO timestamp; clients may pass it but it will be ignored.",
        "properties": {
          "activity_category": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Category"
          },
          "description": {
            "maxLength": 5000,
            "minLength": 20,
            "title": "Description",
            "type": "string"
          },
          "end_time": {
            "format": "date-time",
            "title": "End Time",
            "type": "string"
          },
          "evidence_folder": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Folder"
          },
          "evidence_items": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "title": "Evidence Items",
            "type": "array"
          },
          "evidence_link": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Link"
          },
          "evidence_links": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Links",
            "type": "array"
          },
          "location": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "location_snapshots": {
            "items": {
              "$ref": "#/components/schemas/LocationSnapshot"
            },
            "title": "Location Snapshots",
            "type": "array"
          },
          "material_participation_rentals": {
            "default": false,
            "title": "Material Participation Rentals",
            "type": "boolean"
          },
          "people_involved": {
            "items": {
              "type": "string"
            },
            "title": "People Involved",
            "type": "array"
          },
          "property_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Name"
          },
          "start_time": {
            "format": "date-time",
            "title": "Start Time",
            "type": "string"
          },
          "user": {
            "enum": [
              "Aviv2026",
              "Yarden2026"
            ],
            "title": "User",
            "type": "string"
          }
        },
        "required": [
          "user",
          "description",
          "start_time",
          "end_time"
        ],
        "title": "RepsLogCreate",
        "type": "object"
      },
      "RepsLogRes": {
        "description": "Echoed back to the caller after a successful append.",
        "properties": {
          "activity_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Category"
          },
          "appended_range": {
            "title": "Appended Range",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "end_time": {
            "title": "End Time",
            "type": "string"
          },
          "evidence_folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Folder"
          },
          "evidence_items": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "title": "Evidence Items",
            "type": "array"
          },
          "evidence_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Link"
          },
          "evidence_links": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Links",
            "type": "array"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "location_snapshots": {
            "items": {
              "$ref": "#/components/schemas/LocationSnapshot"
            },
            "title": "Location Snapshots",
            "type": "array"
          },
          "material_participation_rentals": {
            "title": "Material Participation Rentals",
            "type": "boolean"
          },
          "people_involved": {
            "items": {
              "type": "string"
            },
            "title": "People Involved",
            "type": "array"
          },
          "property_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Name"
          },
          "spreadsheet_id": {
            "title": "Spreadsheet Id",
            "type": "string"
          },
          "start_time": {
            "title": "Start Time",
            "type": "string"
          },
          "total_hours": {
            "title": "Total Hours",
            "type": "number"
          },
          "user": {
            "enum": [
              "Aviv2026",
              "Yarden2026"
            ],
            "title": "User",
            "type": "string"
          }
        },
        "required": [
          "created_at",
          "user",
          "property_name",
          "activity_category",
          "description",
          "start_time",
          "end_time",
          "total_hours",
          "location",
          "material_participation_rentals",
          "people_involved",
          "spreadsheet_id",
          "appended_range"
        ],
        "title": "RepsLogRes",
        "type": "object"
      },
...
      "RepsPersonRes": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "name",
          "role",
          "notes"
        ],
        "title": "RepsPersonRes",
        "type": "object"
      },
...
      "RepsPropertyCreate": {
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "RepsPropertyCreate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/reps/activity-categories": {
      "get": {
        "operationId": "reps_list_activity_categories_route_reps_activity_categories_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RepsActivityCategoryRes"
                  },
                  "title": "Response Reps List Activity Categories Route Reps Activity Categories Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Reps List Activity Categories Route"
      },
      "post": {
        "operationId": "reps_create_activity_category_route_reps_activity_categories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepsActivityCategoryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepsActivityCategoryRes"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reps Create Activity Category Route"
      }
    },
    "/reps/activity-categories/{cat_id}": {
      "delete": {
        "operationId": "reps_delete_activity_category_route_reps_activity_categories__cat_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "cat_id",
            "required": true,
            "schema": {
              "title": "Cat Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reps Delete Activity Category Route"
      }
    },
...
    "/reps/entries": {
      "get": {
        "description": "Read the user's full sheet history and return entries + computed stats.",
        "operationId": "reps_entries_route_reps_entries_get",
        "parameters": [
          {
            "in": "query",
            "name": "user",
            "required": true,
            "schema": {
              "title": "User",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepsEntriesEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reps Entries Route"
      }
    },
    "/reps/log": {
      "post": {
        "description": "Append a new REPS entry to the user's Google Sheet (append-only).\n\nEvidence: each `evidence_items` entry becomes a clickable named link in\nthe Sheet's Evidence column (rich text). The user types the label in\nthe modal so the auditor sees `Closing meeting` instead of a 240-char\nGCS URL.\n\nLocation: `location_snapshots` get rendered as breadcrumbs (START/STOP/\nPAUSE/RESUME/BOOKMARK/MANUAL/PHOTO) so an auditor can verify the user\nstayed at the property during the session.",
        "operationId": "reps_log_route_reps_log_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepsLogCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepsLogRes"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reps Log Route"
      }
    },
    "/reps/people": {
      "get": {
        "operationId": "reps_list_people_route_reps_people_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RepsPersonRes"
                  },
                  "title": "Response Reps List People Route Reps People Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Reps List People Route"
      },
      "post": {
        "operationId": "reps_create_person_route_reps_people_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepsPersonCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepsPersonRes"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reps Create Person Route"
      }
    },
    "/reps/people/{person_id}": {
      "delete": {
        "operationId": "reps_delete_person_route_reps_people__person_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reps Delete Person Route"
      },
      "put": {
        "operationId": "reps_update_person_route_reps_people__person_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepsPersonUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepsPersonRes"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reps Update Person Route"
      }
    },
    "/reps/properties": {
      "get": {
        "description": "Bought-deal addresses (priority) + saved prospects.",
        "operationId": "reps_properties_route_reps_properties_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RepsPropertyOption"
                  },
                  "title": "Response Reps Properties Route Reps Properties Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Reps Properties Route"
      },
      "post": {
        "operationId": "reps_create_prospect_route_reps_properties_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepsPropertyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepsPropertyOption"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reps Create Prospect Route"
      }
    },
    "/reps/properties/{prospect_id}": {
      "delete": {
        "operationId": "reps_delete_prospect_route_reps_properties__prospect_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "prospect_id",
            "required": true,
            "schema": {
              "title": "Prospect Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reps Delete Prospect Route"
      }
    },
    "/reps/upload": {
      "post": {
        "description": "Single-file convenience shim \u2014 kept for backward compatibility.\n\nNew clients should call `/reps/upload-batch`.",
        "operationId": "reps_upload_route_reps_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_reps_upload_route_reps_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reps Upload Route"
      }
    },
    "/reps/upload-batch": {
      "post": {
        "description": "Upload one or many evidence files into a per-log GCS sub-folder.\n\nThe folder + per-file URLs are returned to the frontend, which then sends\nthem back inside the `/reps/log` payload so the Sheet stores both the\nfolder URL (auditor's index page) and each file URL.",
        "operationId": "reps_upload_batch_route_reps_upload_batch_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_reps_upload_batch_route_reps_upload_batch_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepsUploadBatchRes"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reps Upload Batch Route"
      }
    },
...
  }
}